reuse urls from sdk and make them a bit safer
Previously we used to have those two constants as regular constants. This is dangerous for whoever is using it, because it uses 'window' and it would break gatsby build just by importing it. Therefore, I've changed it to be a function so that it can be dinamically infered and only requested when on client side.