Feature/generic ext provider (#575)
* generic external providers

* provider name added to auth component props

* Release: 0.278.0-extprov.5

* deleted old repetitive external providers for google and facebook

* fix translations

* nits

* reuse urls from sdk and make them a bit safer

Previously we used to have those two constants as regular constants. This is dangerous for whoever is using it, because it uses 'window' and it would break gatsby build just by importing it. Therefore, I've changed it to be a function so that it can be dinamically infered and only requested when on client side.

* fix flex/box structure

* Release: 0.278.0-extprov.6

* Release: 0.278.0-extprov.7

* Release: 0.278.0-extprov.8

* Remove gitHead

* Release: 0.279.0

Co-authored-by: Larícia Mota <laricia.mota@vtex.com.br>