docs: Fix NextJS and Gatsby sections on using a GraphQL IDE (#1445)
* fix: NextJS and Gatsby getting started steps that were swapped

Only Gatsby provides GraphiQL to explore the API via http://localhost:8000/__graphql. Our NextJS starter doesn't provide something similar, so the section about the using GraphQL Playground is more suitable to it. The section names were inverted.

* fix: A GraphiQL mention instead of GraphQL Playground

But are GraphQL IDEs, but different ones at that.

* fix: NextJS GraphQL API endpont at port 3000 (8000 is Gatsby's)

* chore: Generalize "GraphQL IDE" instead of singling out just GraphiQL

GraphiQL is just for Gatsby so it doesn't make sense to single it out.

* Revert "chore: Generalize "GraphQL IDE" instead of singling out just GraphiQL"

This reverts commit ade518ba8f7dbc45b568d62c570a425be57cc262.

Let's keep this change simple!

* Changing article title

* Changing links

* Update explore-the-faststore-api.mdx

* Update explore-the-faststore-api.mdx

* Update explore-the-faststore-api.mdx

Co-authored-by: PedroAntunesCosta <47991446+PedroAntunesCosta@users.noreply.github.com>