Feat: `Overlay` Component. (#1544)
## What's the purpose of this pull request?

- [x] Move Overlay Component from @faststore/ui to
@faststore/components.
- [x] Adapts docusaurus Overlay documentation.


## How to test it?

run `yarn docusaurus` in app/docs folder and checks the `Overlay`
component or test it using the [doc preview
link](https://faststore-git-feat-fs-ui-overlay-faststore.vercel.app/reference/ui/atoms/Overlay).

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `gatsby.store` AND
`nextjs.store` with this branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->

## References

<!--- Spread the knowledge: is there any content you used to create this
PR that is worth sharing? --->

<!--- Extra tip: adding references to related issues or mentioning
people important to this PR may be good for the documentation and
reviewing process --->

Co-authored-by: Renata Motta <renata.motta@vtex.com.br>