fix: cli copy theme (#1560)
## What's the purpose of this pull request?

When the `src/theme/soft-blue.scss` file does not exist it causes an
error in the console:
```cmd
error - Error: ENOENT: no such file or directory, copyfile '/src/theme/soft-blue.scss' -> '.faststore/src/customizations/themes/index.scss'
```

So I solved that problem and added some validations, like:

When you have the file, but it has not been added to the store.config.js
file, the terminal displays this message:

```cmd
info - The theme needs to be added to the config file to be applied
```

When you don't have the file, but it is added to store.config.js, the
terminal displays this message:

```cmd
info - The soft-blue theme was added to the config file but the soft-blue.scss file does not exist in the themes folder
```

When the file exists and has been added to the configuration, the
terminal displays this message:

```cmd
success - soft-blue theme has been applied
```

## How it works?

I add only a few conditions

## How to test it?

Just use the
[starter.store](https://github.com/vtex-sites/starter.store) repository
for testing.

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `gatsby.store` AND
`nextjs.store` with this branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->

## References

<!--- Spread the knowledge: is there any content you used to create this
PR that is worth sharing? --->

<!--- Extra tip: adding references to related issues or mentioning
people important to this PR may be good for the documentation and
reviewing process --->