feat: new site to documentation (#1561)
## What's the purpose of this pull request?

Create a new documentation site, which is easier to maintain, but also
makes it possible to take advantage of the current documentation that is
written in `.mdx`.

Here is an example of the documentation for the Button component.

<img width="1512" alt="Screenshot 2022-12-17 at 00 13 02"
src="https://user-images.githubusercontent.com/10627086/208221280-cc426caa-9c77-44b5-aad3-aa188044a4ac.png">

## How it works?

The new documentation site is made using [Nextra](https://nextra.site/)

## How to test it?

running `yarn workspace site dev`

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `gatsby.store` AND
`nextjs.store` with this branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->

## References

<!--- Spread the knowledge: is there any content you used to create this
PR that is worth sharing? --->

<!--- Extra tip: adding references to related issues or mentioning
people important to this PR may be good for the documentation and
reviewing process --->