fix: add metadata to search query (#1633)
This PR solves blank intelligent search reports

Returns necessary info so the event can be sent on the front end,
instead of the API. This makes for more accurate events, fired the
correct amount of times with the appropriate data attached to them.

Filter network calls by "sp" and use the search bar on the website. You
should see events with the term and other query information.

https://github.com/vtex-sites/nextjs.store/pull/373