fix(api): add context variables cart mutation (#1659)
## What's the purpose of this pull request?
This PR fixes an inconsistency in the validateCart mutation. If you had
added an item to the cart while it was available for your region, then
you changed to a region where the item would be unavailable, the API
would still return available offers for the item.

This happened because while we were using the Intelligent Search to
fetch the items, we were not passing the right region parameters to it
(because it was not in the context storage). To avoid future
discrepancies, I've also added the locale parameter.

## How it works?

Adds the channel (which contains the region) and locale to the context
storage.

## How to test it?

Go to
https://sfj-9bb2a74--jamstacktata.preview.vtex.app/suprema-de-pollo-en-bandeja-600-g-3883/p
and add it to the cart
Now go to the Locator again and choose "Artigas". On the PDP and
minicart you can see (correctly) that the product's not longer
available.

In production, it'd still be available on the minicart.