Improve resiliency of mergeCMSFile (#1707)
## What's the purpose of this pull request?

Improves the resiliency of `mergeCMSFiles`, by checking if the file CMS
exists before attempting to read it, and showing an error if the JSON
file is malformed.

## How to test it?

On a new fork of
[starter.store](https://github.com/vtex-sites/starter.store), run `yarn
dev` then create an empty `cms` folder. It should break the `dev`.
Restart it and create an empty `cms/faststore/sections.json`. It will
break the `dev` command as well.

Now install the version of this commit, and remove the files created in
the previous steps. Then repeat steps above and you should get no
errors.