feat: add an `user-agent` header to `fetchAPI` method (#1708)
## What's the purpose of this pull request?

<!--- Considering the context, what is the problem we'll solve? Where in
VTEX's big picture our issue fits in? Write a tweet about the context
and the problem itself. --->


Currently `@faststore/api` send requests to VTEX APIs using the default
User Agent of node-fetch package (`node-fetch/${version}`). The purpose
of this requests is to add a custom User Agent to improve observability
of VTEX APIs.

#1709

## How it works?

<!--- Tell us the role of the new feature, or component, in its context.
--->

Change User Agent to `@faststore/api/${version}` in requests.

## How to test it?

<!--- Describe the steps with bullet points. Is there any external link
that can be used to better test it or an example? --->

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `gatsby.store` AND
`nextjs.store` with this branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->

## References

<!--- Spread the knowledge: is there any content you used to create this
PR that is worth sharing? --->

<!--- Extra tip: adding references to related issues or mentioning
people important to this PR may be good for the documentation and
reviewing process --->

---------

Co-authored-by: Emerson Laurentino <emerson@laurentino.co>