Update theme.config.tsx (#1819)
## What's the purpose of this pull request?

Add a banner, considering the content migration to v1.faststore.dev

## How it works?

<!--- Tell us the role of the new feature, or component, in its context.
--->

## How to test it?

<!--- Describe the steps with bullet points. Is there any external link
that can be used to better test it or an example? --->

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `gatsby.store` AND
`nextjs.store` with this branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->

## References

<!--- Spread the knowledge: is there any content you used to create this
PR that is worth sharing? --->

<!--- Extra tip: adding references to related issues or mentioning
people important to this PR may be good for the documentation and
reviewing process --->

---------

Co-authored-by: Mariana Caetano Pereira <67270558+Mariana-Caetano@users.noreply.github.com>