feat: new command to cms sync (#1833)
## What's the purpose of this pull request?

Create a new command on Faststore CLI called `cms-sync` in order to
prevent the user from entering the temporary folder so that the script
can run.

## How it works?

```cmd
faststore cms-sync
```
## How to test it?

running the script above

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `gatsby.store` AND
`nextjs.store` with this branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->

## References

<!--- Spread the knowledge: is there any content you used to create this
PR that is worth sharing? --->

<!--- Extra tip: adding references to related issues or mentioning
people important to this PR may be good for the documentation and
reviewing process --->