fix: maybe sku (#1834)
For some reason when we researched the products we always got the first
item, however this item was not always available, which made it seem
that there was no more item available. So this PR is intended to fix
this problem by taking the item that is available instead of the first
one.