Feat: adds landing page feature (#1841)
## What's the purpose of this pull request?

This Pull Request adds the Landing Page creation feature to faststore

## How it works?

Now the developer can create a new LandingPage in Admin CMS, and this
page will be available in _/slug_ path, based in configs.

## How to test it?

You can to add a new Landing Page in
https://storeframework.myvtex.com/admin/new-cms/faststore/
![Screenshot 2023-06-20 at 18 44
13](https://github.com/vtex/faststore/assets/51174217/b76fd4d7-1d48-4590-9b6e-96f27859e504)

To add sections
![Screenshot 2023-06-20 at 18 47
06](https://github.com/vtex/faststore/assets/51174217/25050d2d-0b3d-4f7a-8500-accd85ccb8c0)

To config the path of this Landing Page

![Screenshot 2023-06-20 at 18 48
19](https://github.com/vtex/faststore/assets/51174217/62785d90-c1a4-4312-b37b-b1e4222a034c)

After publish the new page you will can access this landing page by path

![Screenshot 2023-06-20 at 18 51
58](https://github.com/vtex/faststore/assets/51174217/86a1338b-0c6d-4b66-9def-6e321c2c07ec)

### Alert

he landing pages, PLPs (Product Listing Pages), and Category pages
currently share the same path structure. As a result, when a Landing
Page is created with a path that matches a PLP, the Landing Page takes
priority and the PLP will not be displayed.

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `gatsby.store` AND
`nextjs.store` with this branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->

## References

<!--- Spread the knowledge: is there any content you used to create this
PR that is worth sharing? --->

<!--- Extra tip: adding references to related issues or mentioning
people important to this PR may be good for the documentation and
reviewing process --->

---------

Signed-off-by: Arthur Andrade <arthurfelandrade@gmail.com>