fix: script name (#1869)
When we run the `yarn dev` script on the root, the turbo will run all
`dev` package scripts. But the dev script from the `api` package
modifies the bundle, so he fails when the `core` package tries to run.

So I rename the script in order to avoid modifying the bundle.