chore(ui): add exports to package json (#1924)
## What's the purpose of this pull request?

This PR adds exports definitions to `package.json`. This allows for the
`@faststore/ui` and `@faststore/components` packages to be imported from
Jest tests and be transpiled to CommonJS.

We only included the definition for the ESM files, which were not
sufficient for users importing from CommonJS files. We still export ESM
files as the `main` and `require` exports, which means they'll have to
transpile it to CommonJS, but it will be usable.

I investigated the possibility of exporting these packages as CommonJS
natively as well, but I don't think it's worth it. We'd have to support
it and all kinds of errors could come from this, especially the [dual
package
harzard](https://nodejs.org/api/packages.html#dual-package-hazard).

For using Jest for these packages, enable transpilation by removing them
from the ignored paths (inside your Jest config file):

`transformIgnorePatterns:
["/node_modules/(?!(@faststore/ui|@faststore/components)/)"]`

## How it works?

It now exports definitions of those packages for when they're imported
from CommonJS files.

## References

https://nodejs.org/api/packages.html
https://www.typescriptlang.org/docs/handbook/esm-node.html
https://thesametech.com/how-to-build-typescript-project/