chore: fix package json exports (#1963)
## What's the purpose of this pull request?

This PR has the same purpose as #1924 which was reverted by #1931. Read
it's description to learn more about it.

## How it works?

The difference between this PR and #1924 is that we're now exporting
CommonJS and ESM definitions for the `ui` and components `packages`.
This allows us to define the `main` property on `package.json` pointing
to the CommonJS version, which allows jest to correctly recognize this
dependency without needing to transform it.

If any user of this package uses this through modern js, it is imported
as an ESM package (because of the `module` and `exports` definitions).
The two packages are bundled and tree-shaken by webpack anyway, so it
doesn't affect performance.

I've also experimented using `tsup` to bundle the package, but ended up
facing tree-shaking problems and other configuration errors. I went back
to tsc because of this and for the sake of simplicity.

## How to test it?

See the Starter PR below. If you want to test it in your package or
store, just use the new versions of `core`, `ui` and `components` as
dependency.

### Starters Deploy Preview

https://github.com/vtex-sites/starter.store/pull/155

## References

https://cmdcolin.github.io/posts/2022-05-27-youmaynotneedabundler

And all other dependencies of #1924