fix: search for the entire slug intead of just the first one (#2021)
## What's the purpose of this pull request?

fix the problem, when a page is created in headless cms with more than
one path parameter, '/first/second'

## How it works?

in the getStaticProps function it was changed so that the slug is built
both for the page search in the cms and for the plp

## How to test it?

Just create a new landing page in the cms with more than one parameter
in the path.

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `gatsby.store` AND
`nextjs.store` with this branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->

## References

<!--- Spread the knowledge: is there any content you used to create this
PR that is worth sharing? --->

<!--- Extra tip: adding references to related issues or mentioning
people important to this PR may be good for the documentation and
reviewing process --->