feat: add SDK & select components experimental exports (#2033)
## What's the purpose of this pull request?

This PR fixes an issue where users couldn't import helpful hooks and
components from `@faststore/core` because they were not exported yet.
Most of the hooks were not accessible via `@faststore/sdk`, so users had
no way of doing what they tried to do, such as accessing cart and
session data.

Users could do a few of these hooks themselves, but for the sake of
simplicity, code reuse, and experimentation, they will be made
available.

## How it works?

Import helpful, untested and unoptimized (for this use case, that is,
importing from the starter) hooks & components from
`@faststore/core/experimental`. All exports are appended with
`_unstable` at the end, so it's clear the support is experimental.

To use them, import in a component (override or not) like this:

```tsx
import { useSession_unstable as useSession } from "@faststore/core/experimental";

function MyComponent() {
const session = useSession()
}
```

## How to test it?

Go to this deploy preview and try to use hooks and components exported
by the experimental namespace.

### Starters Deploy Preview

https://github.com/vtex-sites/starter.store/pull/201

## References

I found out that if we export things referenced by the `tsconfig.json`
property `paths`, such as the following, the types for the exported
things wouldn't work, and would be interpreted as any. Moving these
exports to relative path exports fixed the issue.

```tsx
/* Types bugged, says useCart_unstable returns any */
export {
  useCart as useCart_unstable,
  cartStore as cartStore_unstable,
} from 'src/sdk/cart'

/* Types works as expected, useCart_unstable returns correct types */
export {
  useCart as useCart_unstable,
  cartStore as cartStore_unstable,
} from '../../src/sdk/cart'
```